refactor: extract parseUserTime
